% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomcdf.R
\name{binomcdf}
\alias{binomcdf}
\title{Compute the binomcdf function from TI calculators}
\usage{
binomcdf(numtrials, p, x)
}
\arguments{
\item{numtrials}{The number of trials (draws). Example: 5}

\item{p}{The probability of a single success Example: 13/52}

\item{x}{The number of successes. Example: (0,1,2,..)}
}
\value{
the probability
}
\description{
Function from https://github.com/cran/distributions/blob/master/R/binomcdf.R
}
\examples{
library(statshelpR)
binomcdf(5, 13.0/52.0, 0)
}
