% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_z_star.R
\name{calc_z_star}
\alias{calc_z_star}
\title{Calculate the z-star value for a confidence interval}
\usage{
calc_z_star(p)
}
\arguments{
\item{p}{The probability of interest, e.g. 0.95}
}
\value{
The z-score
}
\description{
We assume that the samples of interest are drawn from a Gaussian
distribution. We want to compute the z-star value to compute the
confidence interval for the desired mean'
}
\examples{
z_star <- calc_z_star(0.95)
print(z_star)

}
