% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.star.upper.R
\name{chi2.star.upper}
\alias{chi2.star.upper}
\title{Analyze the variance from a dataset using the chi**2 distribution}
\usage{
chi2.star.upper(n, sigma2.samp, sigma2.pop, up.bound = 70,
  conf.level = 0.95)
}
\arguments{
\item{n}{The number of observations (real number)}

\item{sigma2.samp}{The variance for the sample}

\item{sigma2.pop}{The desired variance for the population. H0 is less than or equal to this.}

\item{conf.level}{The desired confidence level (default 0.95)}

\item{upper.bound}{The upper bounnd for the plot (default 70)}
}
\value{
A dataframe
}
\description{
Analyze the variance from a dataset using the chi**2 distribution
}
