% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_medians.R
\name{col_medians}
\alias{col_medians}
\title{Compute the column medians from a dataframe}
\usage{
col_medians(df, na.rm)
}
\arguments{
\item{df}{A data frame to process}

\item{na.rm}{A boolean. Default: False (remove NA values)}
}
\value{
column medians
}
\description{
Credit: Adapted from Stephen Cohen on twitter
}
\examples{
library(statshelpR)
df <- data.frame(x=c(1.0, 1.5, 2.0, 2.5, 3.0),
                 y=c(2.0, 3.0, 4.0, 5.0, 6.0))
print(col_medians(df, na.rm=FALSE))

}
