% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_devtools.R
\name{ensure_devtools}
\alias{ensure_devtools}
\title{Ensure devtools are installed}
\usage{
ensure_devtools()
}
\description{
Ensure devtools are installed
}
\examples{
library(statshelpR)
ensure_devtools()
}
