% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_new.R
\name{install_new}
\alias{install_new}
\title{Install a new package}
\usage{
install_new(pkg, repo = "http://cran.revolutionanalytics.com/", lib = .Library)
}
\arguments{
\item{pkg}{String. The package to install}

\item{repo}{Sring, Repository. Default: 'http://cran.revolutionanalytics.com/'}

\item{lib}{Library. Default: .Library}
}
\description{
Install a new package
}
\examples{
library(statshelpR)
install_new("scales")
}
