% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_new_github.R
\name{install_new_github}
\alias{install_new_github}
\title{Install new packages from github}
\usage{
install_new_github(pkg, repo)
}
\arguments{
\item{pkg}{String. Package name}

\item{repo}{String. Repository path}
}
\description{
Install new packages from github
}
\examples{
library(statshelpR)
install_new_github('rPeaks','jrminter')
}
