% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olk_data.R
\docType{data}
\name{olk_data}
\alias{olk_data}
\title{Data for Thon Ring analysis}
\format{A data frame with 8 observations of the variables
 \code{n}, code{max}, and \code{s.recip.ang}.}
\source{
Data from Andre Krivanek on p.382 of Spence, High Resolution
 Microscopy, 3rd Ed., Oxford Press (2003).
}
\usage{
data(olk_data)
}
\description{
A dataframe for linear regression analysis. It has the
 following columns: n: the ring number, max: 1 if maximum, 0 if
 minimum, s.recip.ang: the scattering vector in inverse Angstroms.
}
\author{
Andre Krivanek. A standard data set.
}
