% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_sds.R
\name{row_sds}
\alias{row_sds}
\title{Compute the row standard deviations from a dataframe}
\usage{
row_sds(df, na.rm)
}
\arguments{
\item{df}{A data frame to process}

\item{na.rm}{A boolean. Default: False (remove NA values)}
}
\value{
row standard deviations
}
\description{
Credit: Adapted from Stephen Cohen on twitter
}
\examples{
library(statshelpR)
df <- data.frame(x=c(1.0, 1.5, 2.0, 2.5, 3.0),
                 y=c(2.0, 3.0, 4.0, 5.0, 6.0))
print(row_sds(df, na.rm=FALSE))

}
