% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem.leaf.plot.R
\name{stem.leaf.plot}
\alias{stem.leaf.plot}
\title{Plot a Stem and Leaf plot}
\usage{
stem.leaf.plot(dat)
}
\arguments{
\item{dat}{The vector of observations to plot}
}
\value{
A ggplot
}
\description{
Plot a stem and leaf plot from a vector or list of observations
}
\details{
Credit: Adapted from https://stackoverflow.com/questions/26532564/how-to-output-a-stem-and-leaf-plot-as-a-plot
}
