% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_leaf_plot.R
\name{stem_leaf_plot}
\alias{stem_leaf_plot}
\title{Plot a Stem and Leaf plot}
\usage{
stem_leaf_plot(dat)
}
\arguments{
\item{dat}{The vector of observations to plot}
}
\value{
A ggplot
}
\description{
Plot a stem and leaf plot from a vector or list of observations
}
\details{
Credit: Adapted from https://stackoverflow.com/questions/26532564/how-to-output-a-stem-and-leaf-plot-as-a-plot
}
