% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstar_cl_r.R
\docType{data}
\name{tstar_cl_r}
\alias{tstar_cl_r}
\title{The confidence level for t* as a function value for the R qt function}
\format{A data frame with 6 observations of the variable,
 \code{CL} and the variable \code{R}.}
\source{
A table from http://www.stats4stem.org/r-t-distribution
}
\usage{
data(tstar_cl_r)
}
\description{
A dataframe the desired confidence level for finding the
 t-star for a desired confidence level as a function of confidence
 level entered into R's qt function.
}
\author{
STATS4STEM under a Creative Commons licence
}
