% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.interval.R
\name{var.interval}
\alias{var.interval}
\title{Analyze the variance form a dataset using the chi**2 distribution}
\usage{
var.interval(data, sigma0, conf.level = 0.95)
}
\arguments{
\item{data}{A vector of data values (real number)}

\item{sigma0}{The target standard deviation}

\item{conf.level}{The desired confidence level (0.95)}
}
\value{
None. Prints results
}
\description{
Analyze the variance form a dataset using the chi**2 distribution
}
\examples{
library(statshelpR)
# Data from Example 16 page 334 Larsen: Thicknesses of washers
# (in cm) of 15 randomly selected washers are listed below. The
# 95\% confidence interval. The variation is supposed to be within 
# 0.002 cm. Are the washers acceptable?

dat <- c(0.422, 0.424, 0.424, 0.430, 0.419,
         0.424, 0.420, 0.424, 0.425, 0.425,
         0.423, 0.431, 0.437, 0.422, 0.434)

var.interval(dat, 0.002, 0.95)

}
