% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_interval.R
\name{var_interval}
\alias{var_interval}
\title{Analyze the variance from a dataset using the chi**2 distribution}
\usage{
var_interval(data, sigma0, conf_level = 0.95)
}
\arguments{
\item{data}{A vector of data values (real number)}

\item{sigma0}{The target standard deviation}

\item{conf_level}{The desired confidence level (0.95)}
}
\value{
A dataframe with results
}
\description{
This was developed from [eda358.r](https://www.itl.nist.gov/div898/handbook/eda/section3/eda358.r)
I modified the function to return a dataframe. The user can print
the dataframe or use functions like kable to output a table in
a Rmarkdown document. The original function supplied a default
95% confidence interval and let the user change it. However the 
95% CI was hardcoded in the body of the original function.
}
\examples{
library(statshelpR)
# Data from Example 16 page 334 Larsen: Thicknesses of washers
# (in cm) of 15 randomly selected washers are listed below. The
# 95\% confidence interval. The variation is supposed to be within 
# 0.002 cm. Are the washers acceptable?

dat <- c(0.422, 0.424, 0.424, 0.430, 0.419,
         0.424, 0.420, 0.424, 0.425, 0.425,
         0.423, 0.431, 0.437, 0.422, 0.434)

df <- var_interval(dat, 0.002, 0.95)
print(df)

}
